% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_class_function.R
\name{OpenAP}
\alias{OpenAP}
\title{OpenAP Class}
\description{
An R6 class to manage OpenAP functionality, including loading mappings,
individual signal IDs, and signal documentation for a specified release year.
}
\examples{
openap_instance <- OpenAP$new(release_year = 2023)
Initialize the OpenAP Class

Initializes the OpenAP class instance with data for the specified release year (or per default with the latest data).
Loads mappings, individual signal IDs, and signal documentation.

## ------------------------------------------------
## Method `OpenAP$new`
## ------------------------------------------------

openap_instance <- OpenAP$new(release_year = 2023)
List of Available Portfolios

Prints a list of available portfolios in the release.

## ------------------------------------------------
## Method `OpenAP$list_port`
## ------------------------------------------------

openap_instance$list_port()
Get Dataset URL

Retrieves the URL for a specific dataset based on its name.

## ------------------------------------------------
## Method `OpenAP$get_url`
## ------------------------------------------------

openap_instance$get_url("nyse")
Download Portfolio Data

Downloads portfolio data for a specified dataset and optionally filters by predictor.

## ------------------------------------------------
## Method `OpenAP$dl_port`
## ------------------------------------------------

data <- openap_instance$dl_port("deciles_ew", predictor = c("Accruals"))
Get Individual Signal URL

Retrieves the URL for an individual signal based on its name.

## ------------------------------------------------
## Method `OpenAP$get_individual_signal_url`
## ------------------------------------------------

url <- openap_instance$get_individual_signal_url("Accruals.csv")
Download specific Firm Level Characteristics

Downloads specific firm  characteristics

## ------------------------------------------------
## Method `OpenAP$dl_signal`
## ------------------------------------------------

signals <- openap_instance$dl_signal(predictor = c("BM"), signed = TRUE)
Download all Firm Level Characteristics

Downloads all firm level characteristics from the release folder.

## ------------------------------------------------
## Method `OpenAP$dl_all_signals`
## ------------------------------------------------

signals_data <- openap_instance$dl_all_signals()
Download Signal Documentation

Downloads the signal documentation CSV for the release.

## ------------------------------------------------
## Method `OpenAP$dl_signal_doc`
## ------------------------------------------------

signal_doc <- openap_instance$dl_signal_doc()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name_id_map}}{A map of names to IDs.}

\item{\code{individual_signal_id_map}}{A map of individual signal IDs.}

\item{\code{signal_sign}}{The sign of the signal.}

\item{\code{url}}{The base URL for OpenAP operations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OpenAP-new}{\code{OpenAP$new()}}
\item \href{#method-OpenAP-list_port}{\code{OpenAP$list_port()}}
\item \href{#method-OpenAP-get_url}{\code{OpenAP$get_url()}}
\item \href{#method-OpenAP-dl_port}{\code{OpenAP$dl_port()}}
\item \href{#method-OpenAP-get_individual_signal_url}{\code{OpenAP$get_individual_signal_url()}}
\item \href{#method-OpenAP-dl_signal}{\code{OpenAP$dl_signal()}}
\item \href{#method-OpenAP-dl_all_signals}{\code{OpenAP$dl_all_signals()}}
\item \href{#method-OpenAP-dl_signal_doc}{\code{OpenAP$dl_signal_doc()}}
\item \href{#method-OpenAP-clone}{\code{OpenAP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-new"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$new(release_year = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{release_year}}{The release year to initialize (default: latest).}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{openap_instance <- OpenAP$new(release_year = 2023)
List of Available Portfolios

Prints a list of available portfolios in the release.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-list_port"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-list_port}{}}}
\subsection{Method \code{list_port()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$list_port()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{openap_instance$list_port()
Get Dataset URL

Retrieves the URL for a specific dataset based on its name.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-get_url}{}}}
\subsection{Method \code{get_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$get_url(data_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the Portfolio.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A URL string for the dataset.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{openap_instance$get_url("nyse")
Download Portfolio Data

Downloads portfolio data for a specified dataset and optionally filters by predictor.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-dl_port"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-dl_port}{}}}
\subsection{Method \code{dl_port()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$dl_port(data_name, predictor = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the dataset.}

\item{\code{predictor}}{A vector of predictor names to filter (optional).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame containing the portfolio data.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data <- openap_instance$dl_port("deciles_ew", predictor = c("Accruals"))
Get Individual Signal URL

Retrieves the URL for an individual signal based on its name.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-get_individual_signal_url"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-get_individual_signal_url}{}}}
\subsection{Method \code{get_individual_signal_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$get_individual_signal_url(signal_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal_name}}{The name of the signal to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string representing the URL of the signal.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{url <- openap_instance$get_individual_signal_url("Accruals.csv")
Download specific Firm Level Characteristics

Downloads specific firm  characteristics
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-dl_signal"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-dl_signal}{}}}
\subsection{Method \code{dl_signal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$dl_signal(predictor = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{A vector of predictor names to download.}

\item{\code{signed}}{Logical; whether to apply signed transformation based on signal documentation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame containing the signal data.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{signals <- openap_instance$dl_signal(predictor = c("BM"), signed = TRUE)
Download all Firm Level Characteristics

Downloads all firm level characteristics from the release folder.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-dl_all_signals"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-dl_all_signals}{}}}
\subsection{Method \code{dl_all_signals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$dl_all_signals()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame containing all firm level characteristics.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{signals_data <- openap_instance$dl_all_signals()
Download Signal Documentation

Downloads the signal documentation CSV for the release.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-dl_signal_doc"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-dl_signal_doc}{}}}
\subsection{Method \code{dl_signal_doc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$dl_signal_doc()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame containing the signal documentation.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{signal_doc <- openap_instance$dl_signal_doc()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
